fix(examples/nextjs-todo-list): refactor initSupabase
