+++ examples/nextjs-todo-list/.env.local.example
@@ -1,0 +1,3 @@
+# Update these with your Supabase details from your project settings > API
+NEXT_PUBLIC_SUPABASE_URL=https://your-project.supabase.co
+NEXT_PUBLIC_SUPABASE_ANON_KEY=your-anon-key
