--- examples/nextjs-todo-list/lib/api.js
@@ -1,5 +1,0 @@
-import { createClient } from '@supabase/supabase-js'
-import { NEXT_PUBLIC_SUPABASE_URL, NEXT_PUBLIC_SUPABASE_KEY } from './constants'
-
-export const supabase = createClient(NEXT_PUBLIC_SUPABASE_URL, NEXT_PUBLIC_SUPABASE_KEY)
-
